chore: move portals to the portals folder
