+++ apps/meteor/client/views/modal/ModalRegion.tsx
@@ -2,8 +2,8 @@ import { useModal, useCurrentModal } from '@rocket.chat/ui-contexts';
 import type { ReactElement } from 'react';
 import React, { lazy, useCallback } from 'react';
 
-import ModalBackdrop from '../../components/modal/ModalBackdrop';
-import ModalPortal from '../../components/modal/ModalPortal';
+import ModalBackdrop from '../../components/ModalBackdrop';
+import ModalPortal from '../../portals/ModalPortal';
 
 const FocusScope = lazy(() => import('react-aria').then((module) => ({ default: module.FocusScope })));
